moved apiusagechart to component
